refactor(drive): 收敛上传队列生命周期
